Update dependabot config for multiple package ecosystems
Modified the dependabot configuration file to manage dependencies for both GitHub Actions and npm, as well as Composer. The revision instructs Dependabot to run weekly checks for these package ecosystems. The npm configuration specifically includes updates for particular dependencies. Furthermore, the commit messages for npm and Composer updates are prefixed for clarity and scope inclusion.